implemented logic, debugged, organize & commented
Logic that was in progress has been completed, tested, and debugged.

App at stable state, but requires a bit more logic to be added to be completed before styling.

Clicking die changes die state to correct face for produced value, and it adds that value to points for the correct players, and ends turns and reset points to zero upon rolling a value of 1, like it's supposed to.

Clicking the Score Button adds accumulated points to the correct players total score, resets points back to zero, and ends their turn.

Default values apply properly to game when player input is skipped in the pre-game play modal.

Comments added to most logic components describing how input produces the output. Comments also provide destinations and paths for the functions for easy traversing.